+++ src/commands/general/help.js
@@ -8,14 +8,14 @@ module.exports = {
     name: "help",
     aliases: ["commands", "cmd"],
     description: "List all of my commands or info about a specific command.",
-    usage: "(command name)",
+    usage: "(command name / category)",
     //bot_perms: [Permissions.FLAGS.MANAGE_MESSAGES],
     cooldown: 5,
     category: "General",
     async execute(message, args, guildDB, t) {
         const { MessageEmbed } = require("discord.js");
         const beautifyPerms = require("../../functions/beautifyPerms");
-        if (message.channel.type !== "dm") {
+        if (message.channel.type !== "dm" && !args.length) {
             const botPerms = message.guild.me.permissionsIn(message.channel);
             if (!botPerms || !botPerms.has(Permissions.FLAGS.MANAGE_MESSAGES))
                 message.reply(
@@ -25,6 +25,7 @@ module.exports = {
                 );
         }
         const commands = message.client.commands.enabled;
+        const { categories } = message.client;
         const emojiList = {
             first: "⏮",
             back: "⏪",
@@ -36,15 +37,15 @@ module.exports = {
         let pages = [new MessageEmbed()];
         let timeout = 200000; //20 secs timeout
 
-        pages[0].setTitle("Welcome-Bot help");
+        pages[0].setTitle(t("cmds:help.bot-help"));
         if (!args.length) {
             let p;
-            message.client.categories.forEach((cat) => {
+            categories.forEach((cat) => {
                 p = pages.length;
                 let commandsCat = [];
                 pages[p] = new MessageEmbed();
                 pages[p].setTitle(
-                    `Welcome-Bot help - ${t(`categories:${cat.key}`)} Category`
+                    `${t("cmds:help.bot-help")} - ${t(`categories:${cat.key}`)} Category`
                 );
                 message.client.commands.enabled.forEach((command) => {
                     if (command.category === cat.name)
@@ -55,8 +56,8 @@ module.exports = {
                         );
                 });
                 pages[p].addField(
-                    `${cat.emoji} Commands in this category`,
-                    `${commandsCat.join("\n")}`
+                    `${cat.emoji} ${t("cmds:help.in-cat")}`,
+                    `\`\`\`\n${commandsCat.join(" • ")}\n\`\`\``
                 );
             });
             pages[0].setDescription(
@@ -65,7 +66,7 @@ module.exports = {
             pages[0].addField("No of Commands:", `${commands.size}`);
             pages[0].addField(
                 "No of categories:",
-                `${message.client.categories.length}`
+                `${categories.length}`
             );
             pages[0].addField(
                 "Get help for specific command:",
@@ -75,7 +76,7 @@ module.exports = {
                 "What is Cooldown:",
                 "Cooldown is the time that must elapse between each command so that it can be executed again by the user"
             );
-            pages[0].addField("Want list of commands?", "Go to the next page!");
+            pages[0].addField("Commands", `${t("cmds:help.cmds")}`);
 
             const curPage = await message.channel.send({
                 embeds: [
@@ -140,11 +141,17 @@ module.exports = {
         const command =
             commands.get(name) ||
             commands.find((c) => c.aliases && c.aliases.includes(name));
+        const category = categories.find(c => c.name.toLowerCase() === name);
 
-        if (!command) {
-            return message.channel.send(
-                `${t("errors:commandNotFound")}, ${message.author}`
-            );
+        if (!command && !category) {
+            if (!command)
+                return message.channel.send(
+                    `${t("errors:commandNotFound")}, ${message.author}`
+                );
+            if (!category)
+                return message.channel.send(
+                    `${t("errors:categoryNotFound")}, ${message.author}`
+                );
         }
 
         pages[0].setDescription(t(`cmds:help.cmdHelp`, { cmd: command.name }));
@@ -188,7 +195,7 @@ module.exports = {
         if (command.ownerOnly)
             pages[0].addField(
                 "Can be executed by:",
-                "Welcome-Bot owners **ONLY**"
+                "Welcome-Bot developers **ONLY**"
             );
 
         pages[0].addField("Cooldown:", `${command.cooldown || 3} second(s)`);