+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Note that the displayed date is in the format `dd-mm-yyyy`
 ### New features
 
 - New command `disable` - Disable welcome / goodBye logs
+- Support translation
+- New command `hug` - Hug a user
+- New command `cuddle` - Cuddle with a user
+- New command `kiss` - Kiss a user
 
 ### Changes